+++ Data/2_6/Skills/sup_int.lua
@@ -698,6 +698,7 @@ skills["SupportAdditionalXP"] = {
 	requireSkillTypes = { },
 	addSkillTypes = { },
 	excludeSkillTypes = { },
+	supportGemsOnly = true,
 	baseMods = {
 	},
 	qualityMods = {