+++ Data/2_6/Skills/sup_str.lua
@@ -392,6 +392,7 @@ skills["SupportAdditionalLevel"] = {
 	requireSkillTypes = { },
 	addSkillTypes = { },
 	excludeSkillTypes = { },
+	supportGemsOnly = true,
 	baseMods = {
 		mod("ManaCost", "MORE", 25), 
 	},