+++ Modules/CalcOffence-2_6.lua
@@ -293,7 +293,7 @@ function calcs.offence(env, actor)
 	end
 	if skillFlags.trap then
 		local baseSpeed = 1 / modDB:Sum("BASE", skillCfg, "TrapThrowingTime")
-		output.TrapThrowingSpeed = baseSpeed * calcLib.mod(modDB, skillCfg, "TrapThrowingSpeed")
+		output.TrapThrowingSpeed = baseSpeed * calcLib.mod(modDB, skillCfg, "TrapThrowingSpeed") * output.ActionSpeedMod
 		output.TrapThrowingTime = 1 / output.TrapThrowingSpeed
 		if breakdown then
 			breakdown.TrapThrowingTime = { }
@@ -302,6 +302,7 @@ function calcs.offence(env, actor)
 				base = s_format("%.2f ^8(base throwing speed)", baseSpeed),
 				{ "%.2f ^8(increased/reduced throwing speed)", 1 + modDB:Sum("INC", skillCfg, "TrapThrowingSpeed") / 100 },
 				{ "%.2f ^8(more/less throwing speed)", modDB:Sum("MORE", skillCfg, "TrapThrowingSpeed") },
+				{ "%.2f ^8(action speed modifier)",  output.ActionSpeedMod },
 				total = s_format("= %.2f ^8per second", output.TrapThrowingSpeed),
 			})
 		end
@@ -331,7 +332,7 @@ function calcs.offence(env, actor)
 	end
 	if skillFlags.mine then
 		local baseSpeed = 1 / modDB:Sum("BASE", skillCfg, "MineLayingTime")
-		output.MineLayingSpeed = baseSpeed * calcLib.mod(modDB, skillCfg, "MineLayingSpeed")
+		output.MineLayingSpeed = baseSpeed * calcLib.mod(modDB, skillCfg, "MineLayingSpeed") * output.ActionSpeedMod
 		output.MineLayingTime = 1 / output.MineLayingSpeed
 		if breakdown then
 			breakdown.MineLayingTime = { }
@@ -340,6 +341,7 @@ function calcs.offence(env, actor)
 				base = s_format("%.2f ^8(base laying speed)", baseSpeed),
 				{ "%.2f ^8(increased/reduced laying speed)", 1 + modDB:Sum("INC", skillCfg, "MineLayingSpeed") / 100 },
 				{ "%.2f ^8(more/less laying speed)", modDB:Sum("MORE", skillCfg, "MineLayingSpeed") },
+				{ "%.2f ^8(action speed modifier)",  output.ActionSpeedMod },
 				total = s_format("= %.2f ^8per second", output.MineLayingSpeed),
 			})
 		end
@@ -352,7 +354,7 @@ function calcs.offence(env, actor)
 	end
 	if skillFlags.totem then
 		local baseSpeed = 1 / modDB:Sum("BASE", skillCfg, "TotemPlacementTime")
-		output.TotemPlacementSpeed = baseSpeed * calcLib.mod(modDB, skillCfg, "TotemPlacementSpeed")
+		output.TotemPlacementSpeed = baseSpeed * calcLib.mod(modDB, skillCfg, "TotemPlacementSpeed") * output.ActionSpeedMod
 		output.TotemPlacementTime = 1 / output.TotemPlacementSpeed
 		if breakdown then
 			breakdown.TotemPlacementTime = { }
@@ -361,6 +363,7 @@ function calcs.offence(env, actor)
 				base = s_format("%.2f ^8(base placement speed)", baseSpeed),
 				{ "%.2f ^8(increased/reduced placement speed)", 1 + modDB:Sum("INC", skillCfg, "TotemPlacementSpeed") / 100 },
 				{ "%.2f ^8(more/less placement speed)", modDB:Sum("MORE", skillCfg, "TotemPlacementSpeed") },
+				{ "%.2f ^8(action speed modifier)",  output.ActionSpeedMod },
 				total = s_format("= %.2f ^8per second", output.TotemPlacementSpeed),
 			})
 		end
@@ -1635,7 +1638,7 @@ function calcs.offence(env, actor)
 		end
 
 		-- Calculate knockback chance/distance
-		output.KnockbackChance = m_min(100, output.KnockbackChanceOnHit * (1 - output.CritChance / 100) + output.KnockbackChanceOnCrit * output.CritChance / 100)
+		output.KnockbackChance = m_min(100, output.KnockbackChanceOnHit * (1 - output.CritChance / 100) + output.KnockbackChanceOnCrit * output.CritChance / 100 + enemyDB:Sum("BASE", nil, "SelfKnockbackChance"))
 		if output.KnockbackChance > 0 then
 			output.KnockbackDistance = round(4 * calcLib.mod(modDB, cfg, "EnemyKnockbackDistance"))
 			if breakdown then